handle float numbers

+++ new_course/erorr_handling.py
@@ -16,8 +16,8 @@ def divide(a, b):
 
 
 try:
-    num1 = int(input("Enter first number: "))
-    num2 = int(input("Enter second number: "))
+    num1 = float(input("Enter first number: "))
+    num2 = float(input("Enter second number: "))
     print("Result:",divide(num1, num2))
 
 except ValueError:
